+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [2824](https://github.com/influxdb/influxdb/pull/2824) -- Add missing call to WaitGroup.Done in execConn. Thanks @liyichao
 - [2823](https://github.com/influxdb/influxdb/pull/2823) -- Convert OpenTSDB to a service.
 - [2838](https://github.com/influxdb/influxdb/pull/2838) -- Set auto-created retention policy period to infinite.
+- [2829](https://github.com/influxdb/influxdb/pull/2829) -- Re-enable Graphite support as a new Service-style component.
 
 ## v0.9.0-rc32 [2015-06-07]